remove utils folder
