+++ pkg/querier/tripperware/util.go
@@ -2,7 +2,9 @@ package tripperware
 
 import (
 	"context"
+	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"net/http"
+	"sort"
 
 	"github.com/weaveworks/common/httpgrpc"
 
@@ -69,4 +71,24 @@ func DoRequests(ctx context.Context, downstream Handler, reqs []Request, limits 
 	}
 
 	return resps, firstErr
+}
+
+// SliceSamples assumes given samples are sorted by timestamp in ascending order and
+// return a sub slice whose first element's is the smallest timestamp that is strictly
+// bigger than the given minTs. Empty slice is returned if minTs is bigger than all the
+// timestamps in samples.
+func SliceSamples(samples []cortexpb.Sample, minTs int64) []cortexpb.Sample {
+	if len(samples) <= 0 || minTs < samples[0].TimestampMs {
+		return samples
+	}
+
+	if len(samples) > 0 && minTs > samples[len(samples)-1].TimestampMs {
+		return samples[len(samples):]
+	}
+
+	searchResult := sort.Search(len(samples), func(i int) bool {
+		return samples[i].TimestampMs > minTs
+	})
+
+	return samples[searchResult:]
 }
